Fix build, also run tests on Mac (#10)
Fix build when building against CARBON-TO-GITHUB by getting binaries from the clients bin folder (user will have to run update binaries script for the client between building and running the Python tests.).  Tests now run and pass on Mac.

+++ src/Thunkers.cpp
@@ -905,7 +905,7 @@ PyObject* Ballpark::PySetBoid(
 	)
 {
 	ID srcId;
-	long turnOn;
+	int turnOn;
 
 	if (!PyArg_ParseTuple(args, "Li",
 		&srcId,
